feat(api-i18n): enable ttsc
