+++ packages/api-i18n/package.json
@@ -35,6 +35,7 @@
     "jest": "^26.6.3",
     "jest-dynalite": "^3.2.0",
     "rimraf": "^3.0.2",
+    "ttypescript": "^1.5.12",
     "typescript": "^4.1.3"
   },
   "publishConfig": {
@@ -44,11 +45,6 @@
   "scripts": {
     "build": "yarn webiny run build",
     "watch": "yarn webiny run watch"
-  },
-  "adio": {
-    "ignoreDirs": [
-      "src/install"
-    ]
   },
   "gitHead": "8476da73b653c89cc1474d968baf55c1b0ae0e5f"
 }
