+++ packages/api-i18n/src/graphql/crud/locales.crud.ts
@@ -3,7 +3,7 @@ import { TenancyContext } from "@webiny/api-security-tenancy/types";
 import { Context } from "@webiny/handler/types";
 import getPKPrefix from "./utils/getPKPrefix";
 import dbArgs from "./utils/dbArgs";
-import { I18NContext, I18NLocale } from "../../types";
+import { I18NContext, I18NLocale } from "~/types";
 
 export default (context: Context<DbContext, TenancyContext, I18NContext>) => {
     const { db } = context;
