+++ packages/api-i18n/src/graphql/graphql/resolvers/getI18NInformation.ts
@@ -1,5 +1,5 @@
 import { GraphQLFieldResolver } from "@webiny/handler-graphql/types";
-import { I18NContext } from "../../../types";
+import { I18NContext } from "~/types";
 
 const resolver: GraphQLFieldResolver<any, any, I18NContext> = (_, args, context) => {
     const { i18n } = context;
